+++ examples/word-count/src/lib.rs
@@ -2,7 +2,6 @@
 // https://github.com/tildeio/helix-website/blob/master/crates/word_count/src/lib.rs
 
 use pyo3::prelude::*;
-use pyo3::wrap_pyfunction;
 use rayon::prelude::*;
 
 /// Searches for the word, parallelized by rayon