+++ tests/test_exceptions.rs
@@ -1,5 +1,5 @@
 use pyo3::prelude::*;
-use pyo3::{exceptions, py_run, wrap_pyfunction, PyErr, PyResult};
+use pyo3::{exceptions, py_run, PyErr, PyResult};
 use std::error::Error;
 use std::fmt;
 #[cfg(not(target_os = "windows"))]